Fix type of seconds

+++ domain/reporter.py
@@ -4,7 +4,7 @@ from domain.board import Board
 summary = "size: {:3d}, solutions: {:5d}, time: {:8.3f}s, method: {:20s}, last: {}"
 
 
-def report(size: int, seconds: int, name: str, solutions: list[Board]) -> str:
+def report(size: int, seconds: float, name: str, solutions: list[Board]) -> str:
     return summary.format(
         size,
         len(solutions),
